pipeline/pyproject.toml
[tool.poetry]
name = "dwarf"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
scipy = "^1.12.0"
numpy = "^1.26.4"
pandas = "^2.2.1"
scikit-learn = "^1.4.1.post1"
google-cloud-aiplatform = "^1.44.0"
google-cloud-storage = "^2.16.0"
kfp = "^2.7.0"
google-cloud-pipeline-components = "^2.11.0"
black = "24.3.0"
pyarrow = "^15.0.2"
google-cloud-bigquery = {extras = ["pandas"], version = "^3.20.1"}
fsspec = "^2024.3.1"
gcsfs = "^2024.3.1"
pytest = "^8.1.1"
sqlalchemy-bigquery = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
